match: SOURCE/PHILAI -> 15/16 exact, and CurrentHero returns i32
CheckForCreatureUpgrades, CheckBuyStuff, DoAnywhereDDoorTownGate,
SetupRelativeHeroStrengths, GetBestBHC, ValueOfBuyingHero, ValueOfTown,
FightValueOfStack, QuickCombat, HeroInteractionAtTown, BuildHero,
BuildCreature, ComputeUpgradeValue, EvaluateGenericSite, EvaluateHeroEvent.

Mostly x87 operand order and slot names - FightValueOfStack alone needed 26 -
plus two shapes worth naming: EvaluateGenericSite carried a `do {} while (0)`
macro back edge retail does not have, and QuickCombat spells out a redundant
`won == 0 || (won != 0 && defenderHero)` guard rather than the simplification.

The batch's one residual turned out to be a second instance of a family we had
already settled once. playerData::CurrentHero() returns i32, not i8: retail's
`movsbl 0x2(%eax),%ecx` with no temp is only reachable from an accessor
declared to return int, and no cast at the call site reproduces it because the
casts fold. Same class as game::GetPlayerColor, found independently from a
different function - narrow inline accessors in this codebase return int, and
modelling them as i8 spills a temp and grows the frame.

The lane proved that fix and then reverted it because the header was outside
its scope, reporting two full before/after builds instead. Applied here it
lands as measured: TownEvent closes, ADVMGR picks up three more, and nothing
regresses.

SOURCE/PHILAI 49/84 -> 65/84 exact, unit fuzzy 93.00% -> 94.39%; tree
1325 -> 1343, code bytes 45.49% -> 47.28%.

+++ include/SOURCE/playerData.h
@@ -78,7 +78,7 @@ public:
     i32 HasMobileHero(void);
     i32 BuildingsOwned(FactionType, BuildingSlotType, i32);
     i32 NumOfGivenArtifact(ArtifactType);
-    i8 CurrentHero(void) {
+    i32 CurrentHero(void) {
         return m_currentHero;
     }
     i8 CurrentTown(void) {
